requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
Flask_MySQLdb==2.0.0
langchain==0.2.3
langchain_community==0.2.4
langchain_openai==0.1.8
mammoth==1.7.1
openai==1.33.0
pandas==2.2.2
pdfplumber==0.11.0
python-dotenv==1.0.1
python_docx==1.1.2
faiss-cpu==1.8.0